requirements.txt
playwright
playwright-stealth
requests
fastapi
uvicorn
python-dotenv